requirements.txt
Flask == 2.0.3
pandas
imdbpy